Added events service, kibana logs and elasticsearch

+++ tasks-api/routes/task.py
@@ -10,10 +10,13 @@ from database.models import Task
 from PIL import Image
 from io import BytesIO
 
+logger = logging.getLogger("root")
 api_blueprint = Blueprint("task_api", __name__)
+
 
 @api_blueprint.route("/tasks", methods=["GET"])
 def get_tasks():
+    logger.info("Get tasks request")
     tasks = Task.query.all()
     return_tasks = []
     for task in tasks:
@@ -23,12 +26,16 @@ def get_tasks():
 
 @api_blueprint.route("/tasks/<int:id>", methods=["GET"])
 def get_task(id):
-    task = task.query.get(id)
+    logger.info("Get task request")
+    task = Task.query.get(id)
+    if task is None:
+        return "Task not found", 404
     return jsonify(task.serialize())
 
 
 @api_blueprint.route("/tasks", methods=["POST"])
 def create_task():
+    logger.info("Create task request")
     data = request.json
     task = Task(text=data["text"], status=data["status"])
     db.session.add(task)
@@ -38,38 +45,44 @@ def create_task():
 
 @api_blueprint.route("/tasks/<int:id>", methods=["PUT"])
 def update_task(id):
+    logger.info("Update task request")
     data = request.json
     task = Task.query.get(id)
+    if task is None:
+        return "Task not found", 404
+    if "text" in data:
+        task.text = data["text"]
+    if "status" in data:
+        task.status = data["status"]
 
-    if 'text' in data:
-        task.text = data['text']
-    if 'status' in data:
-        task.status = data['status']
-    
     db.session.commit()
     return jsonify(task.serialize())
 
 
 @api_blueprint.route("/tasks/<int:id>", methods=["DELETE"])
 def delete_task(id):
+    logger.info("Delete task request")
     task = Task.query.get(id)
+    if task is None:
+        return "Task not found", 404
     db.session.delete(task)
     db.session.commit()
     return jsonify(task.serialize())
 
+
 # Crie uma rota que recebe uma imagem, retira o fundo e retorna a imagem sem fundo
+
 
 @api_blueprint.route("/tasks/image/remove-background", methods=["POST"])
 def remove_background():
-    logging.info("Remove background request")
+    logger.info("Remove background request")
     data = request.json
-    print(data)
-    if 'image' not in data:
+    if "image" not in data:
         return "Missing image data", 400
 
     # Decode the base64 image
     try:
-        image_data = base64.b64decode(data['image'])
+        image_data = base64.b64decode(data["image"])
         image = Image.open(BytesIO(image_data))
         image = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
     except Exception as e:
@@ -90,7 +103,7 @@ def remove_background():
     cv2.grabCut(image, mask, rect, bgdModel, fgdModel, 5, cv2.GC_INIT_WITH_RECT)
 
     # Modify the mask to get the final result
-    mask2 = np.where((mask == 2) | (mask == 0), 0, 1).astype('uint8')
+    mask2 = np.where((mask == 2) | (mask == 0), 0, 1).astype("uint8")
     image = image * mask2[:, :, np.newaxis]
 
     # Convert back to RGB
@@ -100,4 +113,4 @@ def remove_background():
     img_pil.save(buffer, format="PNG")
     buffer.seek(0)
 
-    return send_file(buffer, mimetype='image/png')
+    return send_file(buffer, mimetype="image/png")
